requirements.txt
beautifulsoup4==4.6.0
numpy==1.14.2
Pillow==5.1.0
requests==2.18.4
scipy==1.0.1
selenium==3.12.0
tensorflow-gpu==1.8.0
